+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add content and style to index ([#73])
 - use celery as queue to run background tasks ([#72])
 - implement web map service client ([#95])
+- add resources (compass and globes as SVG graphics) for map generation ([#101])
 
 ### Bug Fixes
 
@@ -36,3 +37,4 @@
 [#72]: https://github.com/GIScience/sketch-map-tool/pull/72
 [#73]: https://github.com/GIScience/sketch-map-tool/pull/73
 [#95]: https://github.com/GIScience/sketch-map-tool/pull/95
+[#101]: https://github.com/GIScience/sketch-map-tool/pull/101